Backend/be_too/requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
Django==5.2.1
django-celery-beat==2.8.1
django-cors-headers==4.3.1
django-filter==25.1
django-timezone-field==7.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.11
kombu==5.6.2
packaging==25.0
pillow==12.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==7.1.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
stripe==10.2.0
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
